+++ core/src/commonTest/kotlin/at/asitplus/awesn1/CastingTest.kt
@@ -32,11 +32,11 @@ val CastingTest by testSuite {
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asSet() }
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asExplicitlyTagged() }
         shouldThrow<Asn1StructuralException> { Asn1.OctetStringEncapsulating { +Asn1.Null() }.asSequence() }
+        shouldThrow<Asn1StructuralException> {Asn1.OctetStringEncapsulating { +Asn1.Null() }.let { it.asStructure() shouldBe it }}
 
         Asn1.OctetStringEncapsulating { +Asn1.Null() }.let { it.asEncapsulatingOctetString() shouldBe it }
         //Reinterpreting this way must always work
         Asn1.OctetStringEncapsulating { +Asn1.Null() }.let { it.asPrimitive() shouldBe it }
-        Asn1.OctetStringEncapsulating { +Asn1.Null() }.let { it.asStructure() shouldBe it }
     }